+++ cli/src/complete.rs
@@ -761,8 +761,88 @@ pub fn branch_name_equals_any_revision(current: &std::ffi::OsStr) -> Vec<Complet
         .collect()
 }
 
+#[derive(Clone, Debug, PartialEq, Eq)]
+struct PartialFileset<'a> {
+    /// Any prefix before the current path pattern
+    prefix: &'a str,
+    /// The current path pattern being completed
+    current_path: &'a str,
+    /// Whether `current_path` is relative to the repository root
+    is_root: bool,
+}
+
+impl<'a> PartialFileset<'a> {
+    /// Splits the path being completed in `current` from any operators and
+    /// preceding expressions.
+    ///
+    /// Parsing is done on a best-effort basis and may produce incorrect results
+    /// e.g. with quoted names containing characters that look like operators.
+    fn parse(current: &'a std::ffi::OsStr) -> Option<Self> {
+        current.to_str().map(Self::parse_str)
+    }
+
+    fn parse_str(current: &'a str) -> Self {
+        let (token_pos, token) = current
+            .rmatch_indices([':', '~', '|', '&', '(', ',', '"', '\''])
+            .next()
+            .unwrap_or((0, ""));
+        let (prefix, current_path) = match token {
+            // Any space after these tokens should be considered part of the path
+            ":" | "\"" | "'" => current.split_at(token_pos + token.len()),
+            // Other characters are operators that accept whitespace between operands
+            _ => {
+                let name = current[token_pos + token.len()..].trim_ascii_start();
+                current.split_at(current.len() - name.len())
+            }
+        };
+        let is_root = match token {
+            ":" => Self::ends_with_root_pattern_kind(prefix),
+            "\"" | "'" => Self::ends_with_root_pattern_kind(&prefix[..token_pos]),
+            _ => false,
+        };
+
+        PartialFileset {
+            prefix,
+            current_path,
+            is_root,
+        }
+    }
+
+    fn ends_with_root_pattern_kind(prefix: &str) -> bool {
+        const ROOT_PATTERN_KINDS: &[&str] = &["root", "root-file", "root-glob", "root-glob-i"];
+        const MIN_LENGTH: usize = "root:".len();
+
+        if prefix.len() < MIN_LENGTH {
+            return false;
+        }
+        let Some(prefix) = prefix.strip_suffix(':') else {
+            return false;
+        };
+        ROOT_PATTERN_KINDS.iter().any(|kind| {
+            let Some(remaining) = prefix.strip_suffix(kind) else {
+                return false;
+            };
+            match remaining.chars().last() {
+                None => true,
+                Some(c) => !c.is_ascii_alphanumeric() && c != '_' && c != '-',
+            }
+        })
+    }
+
+    fn add_prefix(&self, candidates: Vec<CompletionCandidate>) -> Vec<CompletionCandidate> {
+        if self.prefix.is_empty() {
+            candidates
+        } else {
+            candidates
+                .into_iter()
+                .map(|candidate| candidate.add_prefix(self.prefix))
+                .collect()
+        }
+    }
+}
+
 fn path_completion_candidate_from(
-    current_prefix: &str,
+    fileset: &PartialFileset,
     normalized_prefix_path: &Path,
     path: &Path,
     mode: Option<clap::builder::StyledStr>,
@@ -778,7 +858,7 @@ fn path_completion_candidate_from(
     // Trailing slash might have been normalized away in which case we need to strip
     // the leading slash in the remainder away, or else the slash would appear
     // twice.
-    if current_prefix.ends_with(std::path::is_separator) {
+    if fileset.current_path.ends_with(std::path::is_separator) {
         remainder = remainder.strip_prefix('/').unwrap_or(remainder);
     }
 
@@ -787,7 +867,8 @@ fn path_completion_candidate_from(
         // which `mode` refers.
         Ok(file_completion) => Some(
             CompletionCandidate::new(format!(
-                "{current_prefix}{}",
+                "{}{}",
+                fileset.current_path,
                 file_completion.unwrap_or_default()
             ))
             .help(mode),
@@ -795,25 +876,30 @@ fn path_completion_candidate_from(
 
         // Omit `mode` when completing only up to the next directory.
         Err(mut components) => Some(CompletionCandidate::new(format!(
-            "{current_prefix}{}",
+            "{}{}",
+            fileset.current_path,
             components.next().unwrap()
         ))),
     }
 }
 
-fn current_prefix_to_fileset(current: &str) -> String {
-    let cur_esc = globset::escape(current);
+fn fileset_matching_current_path(fileset: &PartialFileset) -> String {
+    let cur_esc = globset::escape(fileset.current_path);
     let dir_pat = format!("{cur_esc}*/**");
     let path_pat = format!("{cur_esc}*");
-    format!("glob:{dir_pat:?} | glob:{path_pat:?}")
+    let glob = if fileset.is_root { "root-glob" } else { "glob" };
+    format!("{glob}:{dir_pat:?} | {glob}:{path_pat:?}")
 }
 
 fn all_files_from_rev(rev: String, current: &std::ffi::OsStr) -> Vec<CompletionCandidate> {
-    let Some(current) = current.to_str() else {
+    let Some(fileset) = PartialFileset::parse(current) else {
         return Vec::new();
     };
+    fileset.add_prefix(all_files_from_rev_fileset(rev, &fileset))
+}
 
-    let normalized_prefix = normalize_path(Path::new(current));
+fn all_files_from_rev_fileset(rev: String, fileset: &PartialFileset) -> Vec<CompletionCandidate> {
+    let normalized_prefix = normalize_path(Path::new(fileset.current_path));
     let normalized_prefix = slash_path(&normalized_prefix);
 
     with_jj(|jj, _| {
@@ -824,8 +910,12 @@ fn all_files_from_rev(rev: String, current: &std::ffi::OsStr) -> Vec<CompletionC
             .arg("--revision")
             .arg(rev)
             .arg("--template")
-            .arg(r#"path.display() ++ "\n""#)
-            .arg(current_prefix_to_fileset(current))
+            .arg(if fileset.is_root {
+                r#"path ++ "\n""#
+            } else {
+                r#"path.display() ++ "\n""#
+            })
+            .arg(fileset_matching_current_path(fileset))
             .stdout(std::process::Stdio::piped())
             .stderr(std::process::Stdio::null())
             .spawn()
@@ -837,7 +927,7 @@ fn all_files_from_rev(rev: String, current: &std::ffi::OsStr) -> Vec<CompletionC
             .take(1_000)
             .map_while(Result::ok)
             .filter_map(|path| {
-                path_completion_candidate_from(current, &normalized_prefix, Path::new(&path), None)
+                path_completion_candidate_from(fileset, &normalized_prefix, Path::new(&path), None)
             })
             .dedup() // directories may occur multiple times
             .collect())
@@ -846,26 +936,46 @@ fn all_files_from_rev(rev: String, current: &std::ffi::OsStr) -> Vec<CompletionC
 
 fn modified_files_from_rev_with_jj_cmd(
     rev: (String, Option<String>),
-    mut cmd: std::process::Command,
+    cmd: std::process::Command,
     current: &std::ffi::OsStr,
 ) -> Result<Vec<CompletionCandidate>, CommandError> {
-    let Some(current) = current.to_str() else {
+    let Some(fileset) = PartialFileset::parse(current) else {
         return Ok(Vec::new());
     };
+    Ok(
+        fileset.add_prefix(modified_files_from_rev_with_jj_cmd_fileset(
+            rev, cmd, &fileset,
+        )?),
+    )
+}
 
-    let normalized_prefix = normalize_path(Path::new(current));
+fn modified_files_from_rev_with_jj_cmd_fileset(
+    rev: (String, Option<String>),
+    mut cmd: std::process::Command,
+    fileset: &PartialFileset,
+) -> Result<Vec<CompletionCandidate>, CommandError> {
+    let normalized_prefix = normalize_path(Path::new(fileset.current_path));
     let normalized_prefix = slash_path(&normalized_prefix);
 
     // In case of a rename, one entry of `diff` results in two suggestions.
-    let template = indoc! {r#"
-        concat(
-          status ++ ' ' ++ path.display() ++ "\n",
-          if(status == 'renamed', 'renamed.source ' ++ source.path().display() ++ "\n"),
-        )
-    "#};
+    let template = if fileset.is_root {
+        indoc! {r#"
+            concat(
+              status ++ ' ' ++ path ++ "\n",
+              if(status == 'renamed', 'renamed.source ' ++ source.path() ++ "\n"),
+            )
+        "#}
+    } else {
+        indoc! {r#"
+            concat(
+              status ++ ' ' ++ path.display() ++ "\n",
+              if(status == 'renamed', 'renamed.source ' ++ source.path().display() ++ "\n"),
+            )
+        "#}
+    };
     cmd.arg("diff")
         .args(["--template", template])
-        .arg(current_prefix_to_fileset(current));
+        .arg(fileset_matching_current_path(fileset));
     match rev {
         (rev, None) => cmd.arg("--revisions").arg(rev),
         (from, Some(to)) => cmd.arg("--from").arg(from).arg("--to").arg(to),
@@ -890,7 +1000,7 @@ fn modified_files_from_rev_with_jj_cmd(
                 "copied" => "Copied".into(),
                 _ => format!("unknown mode: '{mode}'").into(),
             };
-            path_completion_candidate_from(current, &normalized_prefix, Path::new(path), Some(mode))
+            path_completion_candidate_from(fileset, &normalized_prefix, Path::new(path), Some(mode))
         })
         .collect();
 
@@ -910,11 +1020,17 @@ fn modified_files_from_rev(
 }
 
 fn conflicted_files_from_rev(rev: &str, current: &std::ffi::OsStr) -> Vec<CompletionCandidate> {
-    let Some(current) = current.to_str() else {
+    let Some(fileset) = PartialFileset::parse(current) else {
         return Vec::new();
     };
+    fileset.add_prefix(conflicted_files_from_rev_fileset(rev, &fileset))
+}
 
-    let normalized_prefix = normalize_path(Path::new(current));
+fn conflicted_files_from_rev_fileset(
+    rev: &str,
+    fileset: &PartialFileset,
+) -> Vec<CompletionCandidate> {
+    let normalized_prefix = normalize_path(Path::new(fileset.current_path));
     let normalized_prefix = slash_path(&normalized_prefix);
 
     with_jj(|jj, _| {
@@ -924,7 +1040,7 @@ fn conflicted_files_from_rev(rev: &str, current: &std::ffi::OsStr) -> Vec<Comple
             .arg("--list")
             .arg("--revision")
             .arg(rev)
-            .arg(current_prefix_to_fileset(current))
+            .arg(fileset_matching_current_path(fileset))
             .output()
             .map_err(user_error)?;
         let stdout = String::from_utf8_lossy(&output.stdout);
@@ -937,7 +1053,7 @@ fn conflicted_files_from_rev(rev: &str, current: &std::ffi::OsStr) -> Vec<Comple
                     .next()
                     .expect("resolve --list should contain whitespace after path");
 
-                path_completion_candidate_from(current, &normalized_prefix, Path::new(path), None)
+                path_completion_candidate_from(fileset, &normalized_prefix, Path::new(path), None)
             })
             .dedup() // directories may occur multiple times
             .collect())
@@ -1349,6 +1465,130 @@ mod tests {
     fn test_config_keys() {
         // Just make sure the schema is parsed without failure.
         config_keys();
+    }
+
+    #[test]
+    fn test_partial_fileset_parse_str() {
+        fn partial_fileset<'a>(prefix: &'a str, current_path: &'a str) -> PartialFileset<'a> {
+            PartialFileset {
+                prefix,
+                current_path,
+                is_root: false,
+            }
+        }
+
+        assert_eq!(PartialFileset::parse_str(""), partial_fileset("", ""));
+        assert_eq!(PartialFileset::parse_str(" "), partial_fileset(" ", ""));
+        assert_eq!(PartialFileset::parse_str("foo"), partial_fileset("", "foo"));
+        assert_eq!(
+            PartialFileset::parse_str(" foo"),
+            partial_fileset(" ", "foo")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("glob:"),
+            partial_fileset("glob:", "")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("glob:foo"),
+            partial_fileset("glob:", "foo")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("~foo"),
+            partial_fileset("~", "foo")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("~ foo"),
+            partial_fileset("~ ", "foo")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("glob:*.rs ~ foo"),
+            partial_fileset("glob:*.rs ~ ", "foo")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("foo|bar"),
+            partial_fileset("foo|", "bar")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("foo | bar"),
+            partial_fileset("foo | ", "bar")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("(foo"),
+            partial_fileset("(", "foo")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("( foo"),
+            partial_fileset("( ", "foo")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("function(foo,"),
+            partial_fileset("function(foo,", "")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("function(foo, "),
+            partial_fileset("function(foo, ", "")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("function(foo, b"),
+            partial_fileset("function(foo, ", "b")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("\"foo/b "),
+            partial_fileset("\"", "foo/b ")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("\" foo/b "),
+            partial_fileset("\"", " foo/b ")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("'foo/b "),
+            partial_fileset("'", "foo/b ")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("' foo/b "),
+            partial_fileset("'", " foo/b ")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("function(foo, '"),
+            partial_fileset("function(foo, '", "")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("function(foo, \""),
+            partial_fileset("function(foo, \"", "")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("function(foo, 'b"),
+            partial_fileset("function(foo, '", "b")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("function(foo, \" b"),
+            partial_fileset("function(foo, \"", " b")
+        );
+
+        fn partial_fileset_root<'a>(prefix: &'a str, current_path: &'a str) -> PartialFileset<'a> {
+            PartialFileset {
+                prefix,
+                current_path,
+                is_root: true,
+            }
+        }
+
+        assert_eq!(
+            PartialFileset::parse_str("root:"),
+            partial_fileset_root("root:", "")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("(root-file:foo/"),
+            partial_fileset_root("(root-file:", "foo/")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("~root-glob:'"),
+            partial_fileset_root("~root-glob:'", "")
+        );
+        assert_eq!(
+            PartialFileset::parse_str("~root-glob-i:\"foo"),
+            partial_fileset_root("~root-glob-i:\"", "foo")
+        );
     }
 
     #[test]